Port to upstream gametest.
