+++ Content.IntegrationTests/Fixtures/GameTest.Entities.cs
@@ -211,6 +211,22 @@ public abstract partial class GameTest
     }
 
     /// <summary>
+    ///     Tests if the given entity is deleted on the server.
+    /// </summary>
+    public bool SDeleted(EntityUid target)
+    {
+        return SEntMan.Deleted(target);
+    }
+
+    /// <summary>
+    ///     Tests if the given entity is deleted on the client.
+    /// </summary>
+    public bool CDeleted(EntityUid target)
+    {
+        return CEntMan.Deleted(target);
+    }
+
+    /// <summary>
     ///     Queues an entity for deletion at the end of the tick on the server.
     /// </summary>
     public void SQueueDel(EntityUid id)
@@ -225,4 +241,115 @@ public abstract partial class GameTest
     {
         CEntMan.QueueDeleteEntity(id);
     }
+
+    /// <summary>
+    ///     Queries the number of entities with a given component on the server.
+    /// </summary>
+    public int SQueryCount<T>()
+        where T : IComponent
+    {
+        return Server.EntMan.Count<T>();
+    }
+
+    /// <summary>
+    ///     Queries the number of entities with a given component on the client.
+    /// </summary>
+    public int CQueryCount<T>()
+        where T : IComponent
+    {
+        return Client.EntMan.Count<T>();
+    }
+
+    /// <inheritdoc cref="M:Robust.Shared.GameObjects.EntityManager.EntityQueryEnumerator``1"/>
+    public EntityQueryEnumerator<T> SQuery<T>()
+        where T : IComponent
+    {
+        return Server.EntMan.EntityQueryEnumerator<T>();
+    }
+
+    /// <inheritdoc cref="M:Robust.Shared.GameObjects.EntityManager.EntityQueryEnumerator``1"/>
+    public EntityQueryEnumerator<T> CQuery<T>()
+        where T : IComponent
+    {
+        return Client.EntMan.EntityQueryEnumerator<T>();
+    }
+
+    /// <summary>
+    ///     Gets a single instance of an entity with the given component on the server, asserting it is the only one.
+    /// </summary>
+    public bool SQuerySingle<T>([NotNullWhen(true)] out Entity<T>? ent)
+        where T : IComponent
+    {
+        var query = SQuery<T>();
+
+        if (query.MoveNext(out var eid, out var comp))
+        {
+            Assert.That(query.MoveNext(out var extra, out _),
+                Is.False,
+                $"Expected only one entity with {typeof(T)}, found {SToPrettyString(eid)} and then {SToPrettyString(extra)}");
+            ent = (eid, comp);
+            return true;
+        }
+
+        ent = null;
+        return false;
+    }
+
+    /// <summary>
+    ///     Gets a single instance of an entity with the given component on the client, asserting it is the only one.
+    /// </summary>
+    public bool CQuerySingle<T>([NotNullWhen(true)] out Entity<T>? ent)
+        where T : IComponent
+    {
+        var query = CQuery<T>();
+
+        if (query.MoveNext(out var eid, out var comp))
+        {
+            Assert.That(query.MoveNext(out var extra, out _),
+                Is.False,
+                $"Expected only one entity with {typeof(T)}, found {CToPrettyString(eid)} and then {CToPrettyString(extra)}");
+            ent = (eid, comp);
+            return true;
+        }
+
+        ent = null;
+        return false;
+    }
+
+    /// <summary>
+    ///     Queries every entity with the given component on the server and returns a list of them.
+    /// </summary>
+    public List<Entity<T>> SQueryList<T>()
+        where T : IComponent
+    {
+        var list = new List<Entity<T>>(SQueryCount<T>());
+
+        var q = SQuery<T>();
+
+        while (q.MoveNext(out var ent, out var comp1))
+        {
+            list.Add((ent, comp1));
+        }
+
+        return list;
+    }
+
+    /// <summary>
+    ///     Queries every entity with the given component on the server and returns a list of them.
+    /// </summary>
+    public List<Entity<T>> CQueryList<T>()
+        where T : IComponent
+    {
+        var list = new List<Entity<T>>(CQueryCount<T>());
+
+        var q = CQuery<T>();
+
+        while (q.MoveNext(out var ent, out var comp1))
+        {
+            list.Add((ent, comp1));
+        }
+
+        return list;
+    }
+
 }
