--- Content.IntegrationTests/Tests/_Citadel/Constraints/ConstraintHelpers.cs
@@ -1,63 +1,0 @@
-using System.Diagnostics.CodeAnalysis;
-using Robust.Shared.GameObjects;
-using Robust.Shared.Toolshed.TypeParsers;
-using Robust.UnitTesting;
-
-namespace Content.IntegrationTests.Tests._Citadel.Constraints;
-
-public static class ConstraintHelpers
-{
-    /// <summary>
-    ///     A constraint implementation helper to convert TActual into an entityuid.
-    /// </summary>
-    /// <param name="t">The input value to try to get an entity uid from.</param>
-    /// <param name="ent">The resulting entity uid.</param>
-    /// <param name="error">Whether TActual is recognized to begin with.</param>
-    /// <typeparam name="TActual">The type to cast out of.</typeparam>
-    public static bool TryActualAsEnt<TActual>(TActual t, IIntegrationInstance instance, [NotNullWhen(true)] out EntityUid? ent, out bool error)
-    {
-        if (t is EntityUid u)
-        {
-            ent = u;
-            error = false;
-            return true;
-        }
-
-        if (t is IAsType<EntityUid> asTy)
-        {
-            ent = asTy.AsType();
-            error = false;
-            return true;
-        }
-
-        if (t is IResolvesToEntity resolvable)
-        {
-            if (instance is IServerIntegrationInstance)
-            {
-                ent = resolvable.SEntity;
-            }
-            else if (instance is IClientIntegrationInstance)
-            {
-                ent = resolvable.CEntity;
-            }
-            else
-            {
-                throw new NotSupportedException($"{t.GetType()} is not a valid kind of IIntegrationInstance");
-            }
-
-            error = false;
-            return true;
-        }
-
-        if (t is null)
-        {
-            ent = null;
-            error = false;
-            return false;
-        }
-
-        ent = null;
-        error = true; // Dunno what this type is!
-        return false;
-    }
-}
