--- Content.IntegrationTests/Tests/_Citadel/GameTest.CommonProtos.cs
@@ -1,10 +1,0 @@
-using Content.Shared.Maps;
-using Robust.Shared.Prototypes;
-
-namespace Content.IntegrationTests.Tests._Citadel;
-
-public abstract partial class GameTest
-{
-    // Tiles
-    public static ProtoId<ContentTileDefinition> TilePlating = "Plating";
-}
