--- Content.IntegrationTests/Tests/_Citadel/IResolvesToEntity.cs
@@ -1,10 +1,0 @@
-using Robust.Shared.GameObjects;
-
-namespace Content.IntegrationTests.Tests._Citadel;
-
-public interface IResolvesToEntity
-{
-    EntityUid? SEntity { get; }
-    EntityUid? CEntity { get; }
-}
-
